Larin

+++ src/java/com/mundodental/entidad/Membresias.java
@@ -5,6 +5,7 @@ import com.mundodental.anotaciones.Entity;
 import com.mundodental.anotaciones.FieldName;
 import com.mundodental.anotaciones.NotNull;
 import com.mundodental.anotaciones.PrimaryKey;
+import java.math.BigDecimal;
 import java.sql.Date;
 
 @Entity(table = "Membresias")
@@ -18,14 +19,14 @@ public class Membresias {
     @NotNull
     private Date fechaVencimiento;
     @NotNull
-    private double monto;
+    private BigDecimal monto;
     @NotNull
     private int expediente;
 
     public Membresias() {
     }
 
-    public Membresias(Date fechaRegistro, Date fechaVencimiento, double monto, int expediente) {
+    public Membresias(Date fechaRegistro, Date fechaVencimiento, BigDecimal monto, int expediente) {
         this.fechaRegistro = fechaRegistro;
         this.fechaVencimiento = fechaVencimiento;
         this.monto = monto;
@@ -56,11 +57,11 @@ public class Membresias {
         this.fechaVencimiento = fechaVencimiento;
     }
 
-    public double getMonto() {
+    public BigDecimal getMonto() {
         return monto;
     }
 
-    public void setMonto(double monto) {
+    public void setMonto(BigDecimal monto) {
         this.monto = monto;
     }
